requirements.txt
asgiref==3.7.2
Django==5.0.2
django-appconf==1.0.6
django-cors-headers==4.3.1
django-staticfiles==1.2.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
